close DropdownField on outside click event
