+++ src/components/SelectButton/index.tsx
@@ -172,7 +172,7 @@ export class SelectButton extends React.PureComponent<SelectButtonProps, SelectB
 
     return (
       <RootWrapper {...props}>
-        <ButtonWrapper onMouseDown={this.handleMouseDown} className="ignore-react-onclickoutside">
+        <ButtonWrapper onMouseDown={this.handleMouseDown}>
           {renderValue(data, value)}
           <Icon name="ArrowDropDown" size={1.5} />
         </ButtonWrapper>
